# docs/swagger.yaml
definitions:
...
  auth.LoginRequest:
    properties:
      client_ip:
        type: string
      password:
        maxLength: 72
        minLength: 8
        type: string
      request_id:
        type: string
      request_time:
        type: integer
      user_agent:
        maxLength: 256
        type: string
      username:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - client_ip
    - password
    - request_id
    - request_time
    - username
    type: object
  auth.RegisterRequest:
    properties:
      client_ip:
        type: string
      password:
        maxLength: 72
        minLength: 8
        type: string
      request_id:
        type: string
      request_time:
        type: integer
      user_agent:
        maxLength: 256
        type: string
      username:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - client_ip
    - password
    - request_id
    - request_time
    - username
    type: object
  system.SystemConfig:
    properties:
      app:
        properties:
          uptime:
            type: string
        type: object
      db_host:
        type: string
      db_name:
        type: string
      db_password:
        type: string
      db_port:
        type: integer
      db_user:
        type: string
      debug_mode:
        type: boolean
      environment:
        type: string
      jwt_expire:
        type: integer
      jwt_secret:
        type: string
      log_level:
        type: string
      port:
        type: integer
      redis_db:
        type: integer
      redis_host:
        type: string
      redis_password:
        type: string
      redis_port:
        type: integer
      stats:
        properties:
          cpu_usage_percent:
            type: number
          memory_total:
            type: integer
          memory_used:
            type: integer
          memory_used_percent:
            type: number
        type: object
      system_name:
        type: string
      version:
        type: string
    required:
    - db_host
    - db_name
    - db_password
    - db_port
    - db_user
    - environment
    - jwt_expire
    - jwt_secret
    - log_level
    - port
    - redis_host
    - redis_port
    - system_name
    - version
    type: object
  type_response.BaseResponse:
    properties:
      code:
        description: 业务状态码
        type: integer
      data:
        description: 核心业务数据
      message:
        description: 提示信息
        type: string
      token:
        description: 认证令牌
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /admin/list:
    get:
      consumes:
      - application/json
      description: 分页获取系统中的管理员列表
      parameters:
      - description: 页码
        in: query
        name: page
        required: true
        type: integer
      - description: 每页数量
        in: query
        name: pageSize
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 总数量
          schema:
            type: integer
      summary: 分页获取管理员列表
      tags:
      - Admin
  /admin/login:
    post:
      consumes:
      - application/json
      description: 处理管理员登录请求
      parameters:
      - description: 用户名
        in: body
        name: username
        required: true
        schema:
          type: string
      - description: 密码
        in: body
        name: password
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: 登录成功
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 登录失败
          schema: {}
      summary: 管理员登录
      tags:
      - Admin
  /admin/register:
    post:
      consumes:
      - application/json
      description: 注册新管理员账户
      parameters:
      - description: 用户名
        in: body
        name: username
        required: true
        schema:
          type: string
      - description: 密码
        in: body
        name: password
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: 注册成功
          schema:
            type: string
        "400":
          description: 注册失败
          schema: {}
      summary: 管理员注册
      tags:
      - Admin
  /api/v1/admin/info:
    get:
      consumes:
      - application/json
      description: 获取当前登录管理员的详细信息
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/type_response.BaseResponse'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/type_response.BaseResponse'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/type_response.BaseResponse'
      security:
      - ApiKeyAuth: []
      summary: 获取管理员信息
      tags:
      - 管理员管理
  /api/v1/admin/list:
    get:
      consumes:
      - application/json
      description: 分页获取管理员列表信息
      parameters:
      - default: 1
        description: 页码，默认1
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量，默认10
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/type_response.BaseResponse'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/type_response.BaseResponse'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/type_response.BaseResponse'
      security:
      - ApiKeyAuth: []
      summary: 获取管理员列表
      tags:
      - 管理员管理
  /api/v1/admin/login:
    post:
      consumes:
      - application/json
      description: 处理管理员登录请求，验证用户名和密码，返回JWT令牌
      parameters:
      - description: 登录请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 登录成功
          schema:
            allOf:
            - $ref: '#/definitions/type_response.BaseResponse'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/type_response.BaseResponse'
        "401":
          description: 登录失败
          schema:
            $ref: '#/definitions/type_response.BaseResponse'
      summary: 管理员登录
      tags:
      - 管理员管理
  /api/v1/admin/register:
    post:
      consumes:
      - application/json
      description: 处理管理员注册请求，创建新的管理员账户
      parameters:
      - description: 注册请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 注册成功
          schema:
            allOf:
            - $ref: '#/definitions/type_response.BaseResponse'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/type_response.BaseResponse'
        "409":
          description: 用户已存在
          schema:
            $ref: '#/definitions/type_response.BaseResponse'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/type_response.BaseResponse'
      summary: 管理员注册
      tags:
      - 管理员管理
  /api/v1/admin/update:
    put:
      consumes:
      - application/json
      description: 更新当前登录管理员的信息
      parameters:
      - description: 更新信息参数
        in: body
        name: request
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            allOf:
            - $ref: '#/definitions/type_response.BaseResponse'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/type_response.BaseResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/type_response.BaseResponse'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/type_response.BaseResponse'
      security:
      - ApiKeyAuth: []
      summary: 更新管理员信息
      tags:
      - 管理员管理
  /api/v1/system/config:
    get:
      consumes:
      - application/json
      description: 获取系统配置信息
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/type_response.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/system.SystemConfig'
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取系统配置
      tags:
      - 系统管理
    put:
      consumes:
      - application/json
      description: 更新系统配置信息
      parameters:
      - description: 系统配置信息
        in: body
        name: config
        required: true
        schema:
          $ref: '#/definitions/system.SystemConfig'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/type_response.BaseResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/type_response.BaseResponse'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/type_response.BaseResponse'
      security:
      - ApiKeyAuth: []
      summary: 更新系统配置
      tags:
      - 系统管理
...
  /api/v1/system/info:
    get:
      consumes:
      - application/json
      description: 获取系统基本信息
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/type_response.BaseResponse'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/type_response.BaseResponse'
      security:
      - ApiKeyAuth: []
      summary: 获取系统信息
      tags:
      - 系统管理
...
swagger: "2.0"